requirements.txt
allure-pytest==2.9.44
allure-python-commons==2.9.44
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.4
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
et-xmlfile==1.1.0
html==1.13
idna==3.2
iniconfig==1.1.1
jdcal==1.4.1
openpyexcel==2.5.14
openpyxl==3.0.9
packaging==21.0
PIL-Tools==1.1.0
Pillow==8.3.2
pluggy==1.0.0
py==1.10.0
pyparsing==2.4.7
pytest==6.2.5
requests==2.26.0
selenium==3.141.0
six==1.16.0
toml==0.10.2
urllib3==1.26.6
webdriver-manager==3.4.2
xlrd==2.0.1
